Implement zoom scaling with shortcuts
