+++ src/input/hotkeys.rs
@@ -1,2 +1,5 @@
 /// Input hotkey definitions for PrismX
 pub const SNAP_GRID_TOGGLE: &str = "Ctrl+G";
+pub const ZOOM_IN: &str = "Ctrl+Alt+Z";
+pub const ZOOM_OUT: &str = "Ctrl+Alt+X";
+pub const ZOOM_RESET: &str = "Ctrl+Alt+0";
