+++ src/state.rs
@@ -1,4 +1,4 @@
 // Placeholder file for test plan compatibility.
 // Actual AppState implementation is located in state/mod.rs
-// Fields: debug_input_mode, status_message
+// Fields: debug_input_mode, status_message, zoom_scale
